requirements.txt
requests
pandas
pyarrow
yfinance
tqdm
duckdb
huggingface_hub
alpaca-py
python-dotenv